Set HDX theme

gghdx() gives you the convenience of theme_hdx() without having to explicitly call it for each plot. It also allows for setting the default continuous and discrete scales to follow the HDX color scheme, including default line and point colors and area fills.

gghdx(
  showtext = TRUE,
  base_size = 10,
  base_family = "Source Sans Pro",
  horizontal = TRUE
)

Arguments

showtext

logical If TRUE, uses the showtext package to add the Source Sans Pro font and runs showtext_auto() so all future plots in this session will use the font.

base_size

base font size, given in pts.

base_family

base font family

horizontal

logical Horizontal axis lines?

Details

The default discrete scale is scale_..._hdx() for both fill and color. For continuous scales, the default is scale_fill_gradient_hdx_mint() for fill and scale_color_gradient_hdx_sapphire() for color.