Hex values for HDX colors

hdx_colors() conveniently returns a vector of hex values for specified color ramps. Full values can be found in hdx_color_list. If you know the name of the color you want, such as "sapphire-hdx", you can use hdx_hex(c("sapphire-hdx")) to directly access the hex code.

hdx_colors(colors = c("sapphire", "mint", "tomato", "gray"))

hdx_colours(colors = c("sapphire", "mint", "tomato", "gray"))

hdx_hex(color_names)

hdx_color_names()

hdx_colour_names()

Arguments

colors

Specified color ramps to return. Some set of "sapphire", "mint", "tomato", "gray", "primary", "brand", "neutral", "success", "warning", and "error". By default returns the original "sapphire", "mint", "tomato", and "gray" colors.

color_names

Vector of color names. Valid values are all available using hdx_colors

Value

hdx_colors() returns a named vector of hex values.

hdx_color_names() returns a character vector of color names.

Details

All valid color names are in the named vector returned by hdx_colors() or accessible in the convenient hdx_color_names().

As of the 2025 HDX website redesign, the "primary", "brand", "neutral", "success", "warning", and "error" scales are also available, following the step naming of the HDX design tokens (e.g. "primary-5" for the primary scale's base blue, "neutral-2" for a hairline gridline gray). These are not returned by default to avoid breaking existing uses of hdx_colors(); request them explicitly, e.g. hdx_colors("primary"). The original "sapphire", "mint", and "tomato" scales are kept for backwards compatibility but are superseded by "primary"/"brand"/"error" going forward.

See also

Examples

# get hex values
hdx_colors()
#>        sapphire-dark         sapphire-hdx       sapphire-light 
#>            "#0063B3"            "#007CE0"            "#66B0EC" 
#> sapphire-ultra-light            mint-dark             mint-hdx 
#>            "#CCE5F9"            "#18998F"            "#1EBFB3" 
#>           mint-light     mint-ultra-light          tomato-dark 
#>            "#78D9D1"            "#D2F2F0"            "#C25048" 
#>           tomato-hdx         tomato-light   tomato-ultra-light 
#>            "#F2645A"            "#F7A29C"            "#FCE0DE" 
#>           gray-black            gray-dark          gray-medium 
#>            "#000000"            "#888888"            "#CCCCCC" 
#>           gray-light           gray-white 
#>            "#EEEEEE"            "#FFFFFF" 
hdx_colors("sapphire")
#>        sapphire-dark         sapphire-hdx       sapphire-light 
#>            "#0063B3"            "#007CE0"            "#66B0EC" 
#> sapphire-ultra-light 
#>            "#CCE5F9" 

# 2025 redesign colors
hdx_colors("primary")
#>  primary-0 primary-05  primary-1 primary-15  primary-2  primary-3  primary-4 
#>  "#FFFFFF"  "#E8EFFB"  "#D1E0F7"  "#BAD0F3"  "#A3C0EF"  "#74A1E8"  "#4681E0" 
#>  primary-5  primary-6  primary-7  primary-8 primary-85  primary-9 
#>  "#1862D8"  "#134EAD"  "#0E3B82"  "#0A2756"  "#071D41"  "#05142B" 

# get color names
hdx_color_names()
#>  [1] "tomato-dark"          "tomato-hdx"           "tomato-light"        
#>  [4] "tomato-ultra-light"   "sapphire-dark"        "sapphire-hdx"        
#>  [7] "sapphire-light"       "sapphire-ultra-light" "mint-dark"           
#> [10] "mint-hdx"             "mint-light"           "mint-ultra-light"    
#> [13] "gray-black"           "gray-dark"            "gray-medium"         
#> [16] "gray-light"           "gray-white"           "brand-0"             
#> [19] "brand-05"             "brand-1"              "brand-15"            
#> [22] "brand-2"              "brand-3"              "brand-4"             
#> [25] "brand-5"              "brand-6"              "brand-7"             
#> [28] "brand-8"              "brand-85"             "brand-9"             
#> [31] "brand-95"             "brand-10"             "primary-0"           
#> [34] "primary-05"           "primary-1"            "primary-15"          
#> [37] "primary-2"            "primary-3"            "primary-4"           
#> [40] "primary-5"            "primary-6"            "primary-7"           
#> [43] "primary-8"            "primary-85"           "primary-9"           
#> [46] "neutral-0"            "neutral-01"           "neutral-05"          
#> [49] "neutral-1"            "neutral-15"           "neutral-2"           
#> [52] "neutral-3"            "neutral-4"            "neutral-5"           
#> [55] "neutral-6"            "neutral-7"            "neutral-8"           
#> [58] "neutral-85"           "neutral-9"            "neutral-95"          
#> [61] "neutral-10"           "success-0"            "success-05"          
#> [64] "success-1"            "success-2"            "success-3"           
#> [67] "success-4"            "success-5"            "success-6"           
#> [70] "success-7"            "success-8"            "success-9"           
#> [73] "warning-0"            "warning-05"           "warning-1"           
#> [76] "warning-2"            "warning-3"            "warning-4"           
#> [79] "warning-5"            "warning-6"            "warning-7"           
#> [82] "warning-8"            "warning-9"            "error-0"             
#> [85] "error-05"             "error-1"              "error-2"             
#> [88] "error-3"              "error-4"              "error-5"             
#> [91] "error-6"              "error-7"              "error-8"             
#> [94] "error-9"