Position scales for continuous y data

scale_y_continuous_hdx() and the three variants with different trans arguments are defaults scales for the y axis that ensures the distance from data to the y-axis is reduced to 0, as is common throughout the HDX data visualization guidelines. This is done by setting expand = c(0, 0).

scale_y_continuous_hdx(...)

scale_y_log10_hdx(...)

scale_y_reverse_hdx(...)

scale_y_sqrt_hdx(...)

Arguments

...

Other arguments pass on to ggplot2::scale_y_continuous().

Details

For simple manipulation of labels and limits, you may wish to use labs() and lims() instead.