Hex values for HDX colors

hdx_colors() conveniently returns a vector of hex values for specified color ramps. Full values can be found in hdx_color_list. If you know the name of the color you want, such as "sapphire-hdx", you can use hdx_hex(c("sapphire-hdx")) to directly access the hex code.

hdx_colors(colors = c("sapphire", "mint", "tomato", "gray"))

hdx_colours(colors = c("sapphire", "mint", "tomato", "gray"))

hdx_hex(color_names)

hdx_color_names()

hdx_colour_names()

Arguments

colors

Specified color ramps to return. Some set of "sapphire", "mint", "tomato", and "gray. By default returns all colors.

color_names

Vector of color names. Valid values are all available using hdx_colors

Value

hdx_colors() returns a named vector of hex values.

hdx_color_names() returns a character vector of color names.

Details

All valid color names are in the named vector returned by hdx_colors() or accessible in the convenient hdx_color_names().

See also

Examples

# get hex values
hdx_colors()
#>        sapphire-dark         sapphire-hdx       sapphire-light 
#>            "#0063B3"            "#007CE0"            "#66B0EC" 
#> sapphire-ultra-light            mint-dark             mint-hdx 
#>            "#CCE5F9"            "#18998F"            "#1EBFB3" 
#>           mint-light     mint-ultra-light          tomato-dark 
#>            "#78D9D1"            "#D2F2F0"            "#C25048" 
#>           tomato-hdx         tomato-light   tomato-ultra-light 
#>            "#F2645A"            "#F7A29C"            "#FCE0DE" 
#>           gray-black            gray-dark          gray-medium 
#>            "#000000"            "#888888"            "#CCCCCC" 
#>           gray-light           gray-white 
#>            "#EEEEEE"            "#FFFFFF" 
hdx_colors("sapphire")
#>        sapphire-dark         sapphire-hdx       sapphire-light 
#>            "#0063B3"            "#007CE0"            "#66B0EC" 
#> sapphire-ultra-light 
#>            "#CCE5F9" 

# get color names
hdx_color_names()
#>  [1] "sapphire-dark"        "sapphire-hdx"         "sapphire-light"      
#>  [4] "sapphire-ultra-light" "mint-dark"            "mint-hdx"            
#>  [7] "mint-light"           "mint-ultra-light"     "tomato-dark"         
#> [10] "tomato-hdx"           "tomato-light"         "tomato-ultra-light"  
#> [13] "gray-black"           "gray-dark"            "gray-medium"         
#> [16] "gray-light"           "gray-white"